Change `find` command to be case insensitive
Created new private method `hasIntersection` in FindCommand.java which
will return a boolean value representing whether the two sets passed in
intersect, case insensitively. This is used to compare the `find` query
with the names in the addressbook.

Update tests in FindCommandTest.java to reflect changes, and update the UserGuide.